cdk/emr-serverless-vpc-to-vpc/requirements.txt
aws-cdk-lib==2.43.1
constructs>=10.0.0,<11.0.0
